+++ packages/flutter_tools/test/general.shard/base/logger_test.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 import 'dart:convert' show jsonEncode;
 
+import 'package:flutter_tools/executable.dart';
 import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/base/platform.dart';
